# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PuntosFieles API",
    "version": "1.3.7",
    "description": "Route-derived OpenAPI document for versioned /api/v1 routes, with curated request/response metadata layered onto the discovered route surface."
  },
  "servers": [
    {
      "url": "https://api.puntos.local"
    }
  ],
  "paths": {
...
    "/api/v1/admin/analytics/customer/{id}": {
      "get": {
        "summary": "Get a customer 360 profile for analytics",
        "responses": {
          "200": {
            "description": "Customer analytics profile"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Analytics feature required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-tenant-context": true,
        "x-plan-features": [
          "analytics"
        ]
      }
    },
    "/api/v1/admin/analytics/customer/{id}/ledger": {
      "get": {
        "summary": "Derived GET /api/v1/admin/analytics/customer/{id}/ledger",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-tenant-context": true,
        "x-plan-features": [
          "analytics"
        ]
      }
    },
...
    "/api/v1/admin/analytics/ledger-certification": {
      "get": {
        "summary": "Generate a period-based ledger certification report",
        "responses": {
          "200": {
            "description": "Certification summary and day-level value movement totals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerCertificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Owner role and analytics feature required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive period start. Defaults to 30 days before the end date."
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Inclusive period end. Defaults to today."
          }
        ]
      }
    },
    "/api/v1/admin/analytics/ledger-certification.csv": {
      "get": {
        "summary": "Export the ledger certification report as CSV",
        "responses": {
          "200": {
            "description": "Ledger certification CSV export",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Owner role and analytics feature required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    },
    "/api/v1/admin/analytics/ledger-certification/archive": {
      "get": {
        "summary": "Derived GET /api/v1/admin/analytics/ledger-certification/archive",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "analytics"
        ]
      }
    },
    "/api/v1/admin/analytics/ledger-certification/archive/{date}": {
      "get": {
        "summary": "Derived GET /api/v1/admin/analytics/ledger-certification/archive/{date}",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "analytics"
        ]
      }
    },
    "/api/v1/admin/analytics/ledger-certification/archive/{date}.csv": {
      "get": {
        "summary": "Derived GET /api/v1/admin/analytics/ledger-certification/archive/{date}.csv",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "analytics"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "staffAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_staff"
      },
      "customerAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_customer"
      },
      "superAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_super"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "LedgerCertificationResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "business": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "slug"
            ],
            "additionalProperties": false
          },
          "generated_at": {
            "type": "string"
          },
          "period": {
            "type": "object",
            "properties": {
              "from": {
                "type": "string"
              },
              "to": {
                "type": "string"
              }
            },
            "required": [
              "from",
              "to"
            ],
            "additionalProperties": false
          },
          "certification_status": {
            "type": "string",
            "enum": [
              "OK",
              "REVIEW_REQUIRED"
            ]
          },
          "summary": {
            "type": "object",
            "properties": {
              "points_issued": {
                "type": "number"
              },
              "points_redeemed": {
                "type": "number"
              },
              "points_reversed": {
                "type": "number"
              },
              "points_expired": {
                "type": "number"
              },
              "adjustment_points": {
                "type": "number"
              },
              "gift_cards_issued_q": {
                "type": "number"
              },
              "gift_cards_redeemed_q": {
                "type": "number"
              },
              "replay_events": {
                "type": "number"
              },
              "pending_corrections_count": {
                "type": "number"
              },
              "negative_balance_count": {
                "type": "number"
              },
              "latest_reconciliation_mismatches": {
                "type": "number"
              },
              "latest_reconciliation_completed_at": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "points_issued",
              "points_redeemed",
              "points_reversed",
              "points_expired",
              "adjustment_points",
              "gift_cards_issued_q",
              "gift_cards_redeemed_q",
              "replay_events",
              "pending_corrections_count",
              "negative_balance_count",
              "latest_reconciliation_mismatches",
              "latest_reconciliation_completed_at"
            ],
            "additionalProperties": false
          },
          "daily_rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string"
                },
                "points_issued": {
                  "type": "number"
                },
                "points_redeemed": {
                  "type": "number"
                },
                "points_reversed": {
                  "type": "number"
                },
                "points_expired": {
                  "type": "number"
                },
                "adjustment_points": {
                  "type": "number"
                },
                "gift_cards_issued_q": {
                  "type": "number"
                },
                "gift_cards_redeemed_q": {
                  "type": "number"
                },
                "replay_events": {
                  "type": "number"
                }
              },
              "required": [
                "date",
                "points_issued",
                "points_redeemed",
                "points_reversed",
                "points_expired",
                "adjustment_points",
                "gift_cards_issued_q",
                "gift_cards_redeemed_q",
                "replay_events"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "ok",
          "business",
          "generated_at",
          "period",
          "certification_status",
          "summary",
          "daily_rows"
        ],
        "additionalProperties": true
      },
...
    }
  }
}